+++ .gitignore
@@ -115,3 +115,4 @@ assets/**/dist
 # Gradle
 .idea/**/gradle.xml
 .idea/**/libraries
+cdk.out
