+++ cdk.json
@@ -1,3 +1,3 @@
 {
-    "app": "node index"
+    "app": "npx ts-node cdk/deployment.ts"
 }
