+++ cdk/settings.json
@@ -1,0 +1,4 @@
+{
+    "repositoryName": "aws-sandbox",
+    "environment": "dev"
+}
