+++ test/cdk.unit.test.ts
@@ -1,11 +1,11 @@
 import { expect as expectCDK, exactlyMatchTemplate } from '@aws-cdk/assert';
 import * as cdk from '@aws-cdk/core';
-import * as sut from '../cdk/dynamoDBStack';
+import * as sut from '../cdk/deployment';
 
 test('exact match (case 1)', () => {
     const app = new cdk.App();
     // WHEN
-    const stack = new sut.DynamoDBStack(app, 'MyTestStack');
+    const stack = new sut.Deployment(app, 'MyTestStack');
 
     // Case 1 - exact match
     expectCDK(stack).to(
@@ -16,21 +16,16 @@ test('exact match (case 1)', () => {
                     Properties: {
                         KeySchema: [
                             {
-                                AttributeName: 'itemId',
+                                AttributeName: 'id',
                                 KeyType: 'HASH',
                             },
                         ],
                         AttributeDefinitions: [
                             {
-                                AttributeName: 'itemId',
+                                AttributeName: 'id',
                                 AttributeType: 'S',
                             },
                         ],
-                        ProvisionedThroughput: {
-                            ReadCapacityUnits: 5,
-                            WriteCapacityUnits: 5,
-                        },
-                        TableName: 'items',
                     },
                     UpdateReplacePolicy: 'Delete',
                     DeletionPolicy: 'Delete',
