+++ tsconfig.json
@@ -66,7 +66,9 @@
 
         /* Advanced Options */
         "skipLibCheck": true /* Skip type checking of declaration files. */,
-        "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
+        "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */,
+
+        "resolveJsonModule": true
     },
     "exclude": ["./assets/"]
 }
